Add structured `Events`s to `IndexTx.events`
