+++ packages/cosmwasm-stargate/src/cosmwasmclient.ts
@@ -10,6 +10,7 @@ import {
   BroadcastTxError,
   Coin,
   DeliverTxResponse,
+  fromTendermint34Event,
   IndexedTx,
   isSearchByHeightQuery,
   isSearchBySentFromOrToQuery,
@@ -462,6 +463,7 @@ export class CosmWasmClient {
         height: tx.height,
         hash: toHex(tx.hash).toUpperCase(),
         code: tx.result.code,
+        events: tx.result.events.map(fromTendermint34Event),
         rawLog: tx.result.log || "",
         tx: tx.tx,
         gasUsed: tx.result.gasUsed,
